Lazy loading for GenericErrorView.kt
