+++ moderatorFrontend/src/jsMain/kotlin/views/common/GenericErrorViewConfig.kt
@@ -1,0 +1,7 @@
+package views.common
+
+// TODO: @mh This makes the GenericErrorView.kt lazy component fail if it's in the same file.
+class GenericErrorViewConfig(
+  val title: String = "",
+  val subtitle: String = "",
+)
