+++ moderatorFrontend/src/jsMain/kotlin/views/common/NetworkErrorView.kt
@@ -1,0 +1,17 @@
+package views.common
+
+import react.ChildrenBuilder
+import react.Suspense
+import util.Strings
+import util.get
+
+fun ChildrenBuilder.networkErrorView() {
+  Suspense {
+    GenericErrorViewFc {
+      config = GenericErrorViewConfig(
+        title = Strings.network_error.get(),
+        subtitle = Strings.network_error_description.get(),
+      )
+    }
+  }
+}
